chore: rename some fe api request files and shuffled some functions
