+++ frontend/src/pages/Logout/Logout.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useQueryClient } from '@tanstack/react-query';
-import { logout } from 'utils/api/auth';
+import { logout } from 'utils/api/authApi';
 import PageLoading from 'components/PageLoading';
 import { useAppDispatch, useAppSelector } from 'hooks';
 import { selectToken, unsetIdentity } from 'reducers/identitySlice';
