+++ frontend/src/utils/api/timetableApi.ts
